Move non-trivial things out of the engine package object, move stuff only used by backend to backend package object
