+++ src/main/scala/cromwell/engine/backend/package.scala
@@ -1,0 +1,17 @@
+package cromwell.engine
+
+import scala.concurrent.{ExecutionContext, Future}
+
+package object backend {
+  implicit class EnhancedFutureFuture[A](val ffa: Future[Future[A]])(implicit ec: ExecutionContext) {
+    def flatten: Future[A] = ffa flatMap { fa => fa }
+  }
+
+  implicit class EnhancedExecutionHandle(val handle: ExecutionHandle) extends AnyVal {
+    def future = Future.successful(handle)
+  }
+
+  implicit class EnhancedExecutionResult(val result: ExecutionResult) extends AnyVal {
+    def future = Future.successful(result)
+  }
+}
